split vcpuos.h into two files so the base class is seperate
