+++ src/sst/elements/vanadis/vanadis.cc
@@ -16,8 +16,6 @@
 #include <sst_config.h>
 #include "vanadis.h"
 
-#include "decoder/vmipsdecoder.h"
-#include "decoder/vriscv64decoder.h"
 #include "inst/vinstall.h"
 #include "velf/velfinfo.h"
