+++ src/sst/elements/vanadis/vfpflags.h
@@ -21,7 +21,7 @@ namespace Vanadis {
 
 enum class VanadisFPRoundingMode { ROUND_NEAREST, ROUND_TO_ZERO, ROUND_DOWN, ROUND_UP, ROUND_NEAREST_TO_MAX };
 
-uint64_t convertRoundingToInteger(VanadisFPRoundingMode mode) {
+inline uint64_t convertRoundingToInteger(VanadisFPRoundingMode mode) {
 	switch(mode) {
 	case VanadisFPRoundingMode::ROUND_NEAREST:
 		return 0;
